Add files via upload
Reddit NLP Topic Modeling Prediction
This notebook contains a machine learning model for predicting the topic of Reddit posts based on their title and content. The model uses Natural Language Processing (NLP) techniques and Topic Modeling to categorize posts into one of six topics: Technology, Politics, Science, Business, Sports, and Entertainment.

Data
The dataset used for this project was sourced from Reddit and contains over 1.7 million posts from various subreddits. The dataset was preprocessed to remove irrelevant information such as URLs and punctuation.

Methodology
The model uses a combination of CountVectorizer and Latent Dirichlet Allocation (LDA) to extract relevant features from the text data and identify topics. The model was trained on a subset of the data and achieved an accuracy score of 78%.

Usage
To use the model, simply input a title and content of a Reddit post and run the prediction function. The model will output the most likely topic for the post.

Dependencies
This notebook requires Python 3 and the following libraries:

Pandas
Numpy
Scikit-learn
Gensim
NLTK